requirements.txt
asyncio
asyncpg
aiosqlite
aiohttp
asyncio
discord.py==1.7.3